+++ src/main/java/bananacore/epic/interfaces/OdometerInterface.java
@@ -3,5 +3,5 @@ package bananacore.epic.interfaces;
 import java.sql.Timestamp;
 
 public interface OdometerInterface {//
-    public void updateOdometer(int value, Timestamp timestamp);
+    public void updateOdometer(double value, Timestamp timestamp);
 }
